requirements.txt
brew==0.1.3
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
ffmpeg==1.4
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.2
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
pycairo==1.27.0
PyGObject==3.50.0
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yadisk==3.1.0
